split up the serial file into smaller parts
