+++ web_interface_vue/tamagometer/src/components/AppButtonRequestSerial.vue
@@ -1,13 +1,14 @@
 <script setup lang="ts">
 import { portNeedsToBeRequested } from '@/state';
-import { connectToPort } from '@/serial';
+import { connectToPort } from '@/serialworkerinterface';
 
 
 </script>
 
 <template>
     <div v-if="portNeedsToBeRequested" class="request-port-container">
-        <button class="icon-label-button" @click="connectToPort(true).then(success => portNeedsToBeRequested = !success)">
+        <button class="icon-label-button"
+            @click="connectToPort(true).then(success => portNeedsToBeRequested = !success)">
             <svg class="round-button-icon" transform="rotate(90)" viewBox="0 0 80 80" fill="none"
                 xmlns="http://www.w3.org/2000/svg">
                 <path d="M26 48L26 58C26 65.732 32.268 72 40 72C47.732 72 54 65.732 54 58V48" />
